re #18: Refactored pages to use a common show?entityType=... form
This replaces the page for each type of entity that was needlessly
repetitive.

The sitemap is now updated to use a more compact definition form that
does not require specifying every allowed path.

The common page also has the page title and add button on the top row
which is probably cleaner.