closes #57 and closes #41: Fixed issue where supplier details were not updated when a new part was selected in the Line Item entry.
This was essentially caused by the SHtml.Replace command not setting
the id attribute in the new html nodes that were generated so subsequent
calls could not find an element with the given id and so never
updated the page.

This also allowed me to enable the reset of part, supplier and quantity
values on successful adding of a new line.