re #73: Exposed `validate(ValidationItem*)` method in `DataValidation` to enable `PartCostSnippet` to use validation.
The problem here was due to the fact that the `PartCostSnippet` is a
trait and the owning snippet also defined a `validationItems` override
which was called when the `PartCostSnippet` called `performValidation`.
This caused the part cost adding to call the validation for the supplier
(checking address etc) and not for the part cost entry.