re #68: Refactoring of `AddressParser` and update of ScalaDoc
The `AddressParser` now takes a list of address lines as input rather
than a flattened string. These are then stripped of trailing commas and
full stops, trimmed and empty lines dropped. This is then used to
generate the address if there is a Country Code match and there is at
least one address line.