re #35: Removed adding/modification of `Address` records from the `add` and `modify` methods of `Customer`
Also changed tests that checked the addition of `Address` records to
check that the `Address` records are not added or modified when updating
the `Customer` record.

This brings the `Customer` class in line with the rest of the Record
classes that assume that any fields which use `ObjectIdRef`'s are
handled externally to the record that references them.