close #95: Removed all use of forcing https on CF.
This is basically due to the fact that CF terminates all https at their
boundary and lift sees any requests as http so any attempt to check the
request scheme will always see http.

The only solution I can see now is to ensure that the bookmark/link
used is an https one as this seems to propagate correctly across the
session.

I suppose it is also possible to use a session/request var to count the
redirects and what I think the scheme is, but not for the moment.